requirements.txt
# Ubik Project Requirements
# PowerSpec AI100 - RTX 5090 - CUDA 12.4
# Python 3.12

# ============================================================================
# Core ML Framework
# ============================================================================
torch>=2.9.1
torchvision>=0.24.0
torchaudio>=2.9.0

# ============================================================================
# Transformer Models & Inference
# ============================================================================
transformers>=4.57.0
accelerate>=1.12.0
sentencepiece>=0.2.0
protobuf>=4.0.0
safetensors>=0.4.0

# ============================================================================
# Quantization & Memory Optimization
# ============================================================================
bitsandbytes>=0.49.0
optimum>=1.22.0

# ============================================================================
# Training & Fine-tuning
# ============================================================================
peft>=0.14.0          # LoRA, QLoRA adapters
trl>=0.14.0           # DPO, RLHF training
datasets>=4.0.0       # HuggingFace datasets
evaluate>=0.4.0       # Model evaluation metrics

# ============================================================================
# Utilities & Infrastructure
# ============================================================================
huggingface-hub>=0.36.0
tokenizers>=0.21.0
einops>=0.8.0
wandb>=0.19.0         # Experiment tracking (optional)
tensorboard>=2.18.0   # TensorBoard logging

# ============================================================================
# Development & Analysis
# ============================================================================
jupyter>=1.1.0
jupyterlab>=4.0.0
ipython>=8.20.0
matplotlib>=3.9.0
seaborn>=0.13.0
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.14.0

# ============================================================================
# API & Web Services
# ============================================================================
fastapi>=0.115.0      # API server
uvicorn>=0.34.0       # ASGI server
pydantic>=2.10.0      # Data validation
aiohttp>=3.11.0       # Async HTTP client

# ============================================================================
# MCP (Model Context Protocol) - for Cognitive Node communication
# ============================================================================
# mcp>=1.0.0          # Uncomment when available
httpx>=0.28.0         # HTTP client for MCP
websockets>=14.0      # WebSocket support
python-dotenv>=1.0.0  # Environment variables

# ============================================================================
# Monitoring & Logging
# ============================================================================
psutil>=6.1.0         # System monitoring
gpustat>=1.1.0        # GPU monitoring
tqdm>=4.67.0          # Progress bars
loguru>=0.7.0         # Better logging

# ============================================================================
# Testing
# ============================================================================
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0

